% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{prettier_cols}
\alias{prettier_cols}
\title{Function to extract pretty colors as hex codes}
\usage{
prettier_cols(...)
}
\arguments{
\item{...}{Character names of pretty_colors}
}
\description{
source:
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
