% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{prettier_pal}
\alias{prettier_pal}
\title{Return function to interpolate a prettier color palette}
\usage{
prettier_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in prettier_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a prettier color palette
}
