% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.R
\name{save_data}
\alias{save_data}
\title{Saves data underlying a ggplot}
\usage{
save_data(filename, plot = last_plot())
}
\arguments{
\item{filename}{Base file name/path}

\item{plot}{= last_plot() The ggplot for which the data will be saved}
}
\value{
a file is saved
}
\description{
Helper function that saves the data used to create a gglot.
This should help reproducing plots.
}
