% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.R
\name{save_poster}
\alias{save_poster}
\title{Saves plots}
\usage{
save_poster(plot, filename, width = 15, height = 15, units = "cm")
}
\arguments{
\item{plot}{The gggplot2 object to be saved}

\item{filename}{Base file name/path}

\item{width}{width of the save image. Default: 15 cm}

\item{height}{height of the save image. Default: 15 cm}
}
\value{
A file is saved
}
\description{
Helper function that saves (gg)plots as PNG and PDF simultaneously
The default size should save plots large enough for a poster
Modified from: https://gist.github.com/sheymann/2399659
}
