% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_prettier}
\alias{scale_color_prettier}
\title{Color scale constructor for prettier colors}
\usage{
scale_color_prettier(palette = "main", ...)
}
\arguments{
\item{palette}{Character name of palette in prettier_palettes}

\item{...}{Unused for now}
}
\description{
Color scale constructor for prettier colors
}
