% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_colour_ketting}
\alias{scale_colour_ketting}
\title{Gives some color to plots}
\usage{
scale_colour_ketting(...)
}
\value{
A color scale for ggplot
}
\description{
A color scheme for plots that is both colorblind-friendly
and yet is still aesthetically pleasing
}
