% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_prettier}
\alias{scale_fill_prettier}
\title{Fill scale constructor for prettier colors}
\usage{
scale_fill_prettier(palette = "main", ...)
}
\arguments{
\item{palette}{Character name of palette in prettier_palettes}

\item{...}{Unused for now}
}
\description{
Fill scale constructor for prettier colors
}
