% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_poster.R
\name{theme_poster}
\alias{theme_poster}
\title{Theme optimized for poster presentations}
\usage{
theme_poster(base_size = 16, base_family = "Helvetica")
}
\arguments{
\item{base_size}{The font size. Default 16}

\item{base_family}{The font type. Default "Helvetica"}
}
\value{
A gggplot theme function
}
\description{
Colors sizes, legends and in particular font sizes have been optimized
for posters. 
When saving, a size of 10 cm will make sure that very little editing is
needed in illustrator.
}
